Add basic h5py and sqlite prototypes
Includes code to convert from BIOM Format 1.0.0 to these formats, as well as code to parse the new formats into Table objects. The Table objects have some very basic operations implemented.

This code was copied over from the 'new-format' branch of my fork of the biom-format repo (https://github.com/jrrideout/biom-format/tree/new-format). I think this will be a better home for it as it currently isn't tied to anything in the biom-format repo and it is purely prototype code. Development will stop on the 'new-format' branch in favor of this repo.